refactor(layout): extract canonical render pipeline into sdpm.layout
Extract cmd_layout's inline pipeline into reusable pure functions so the
CLI, the QA harness, and (future) MCP tools all call one source of truth
instead of re-implementing scale/route/measure.

- Add sdpm/layout/render.py: build_layout() + render_architecture() ->
  {elements, bbox, warnings, metrics}. Respects the root `reverse` flag.
- Add sdpm/layout/metrics.py: measure/measure_layout/score, moved from
  layout_qa.py.
- Move _segments_cross into sdpm.layout (builder re-exports it).
- cmd_layout is now a thin wrapper (include_metrics=False keeps stdout
  byte-identical to prior versions).
- layout_qa.py drops its duplicated _build_layout and delegates to the
  canonical pipeline; still re-exports measure/score for layout_search.py.

Fixes a latent bug: layout_qa's duplicate pipeline ignored `reverse`, so
QA metrics disagreed with what the CLI actually drew for reversed diagrams.

Verified against HEAD across 7 diverse diagrams: CLI stdout byte-identical;
QA metrics identical except the one reversed diagram (wire_norm/score now
reflect true geometry). ruff clean, 188 tests pass.
